requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
fastapi
uvicorn
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
matplotlib
apscheduler